requirements.txt
aioboto3~=15.0.0
aiodataloader~=0.4.2
aiodocker==0.24.0
aiofiles~=24.1.0
aiohttp~=3.13.3
aiohttp_cors~=0.8.1
aiohttp_jinja2~=1.6
aiohttp_sse>=2.2
aiodns==3.2
aiomonitor~=0.7.0
aioresponses>=0.7.3
aiosqlite~=0.21.0
aiotools~=2.2.3
aiotusclient~=0.1.4
alembic~=1.13.2
appdirs~=1.4.4
async_timeout~=4.0
asyncpg>=0.29.0
asynctest>=0.13.0
asyncudp>=0.11
attrs>=25.3
bcrypt~=4.2.0
boto3~=1.35
cachetools~=5.5.0
callosum~=1.0.3
cattrs~=24.1.1
click~=8.1.7
coloredlogs~=15.0
colorama>=0.4.6
cryptography>=44.0.2
dataclasses-json~=0.5.7
faker~=24.7.1
graphene~=3.3.0
graphene-federation~=3.2.0
graphql-core~=3.2.6
graphql-relay~=3.2.0
graypy==2.1.0
humanize>=3.1.0
ifaddr~=0.2
inquirer~=3.3.0
janus~=2.0
Jinja2~=3.1.6
jupyter-client>=8.6
kubernetes~=33.1.0
kubernetes-asyncio~=33.3.0
lark~=1.1.5
memray~=1.17.2
more-itertools~=10.5.0
msgpack~=1.1.0
multidict~=6.6.4
namedlist~=1.8
networkx~=3.3.0
orjson~=3.10.16
opentelemetry-api~=1.39.0
opentelemetry-sdk~=1.39.0
opentelemetry-exporter-otlp-proto-grpc~=1.39.0
opentelemetry-instrumentation-aiohttp-client~=0.60b0
opentelemetry-instrumentation-aiohttp-server~=0.60b0
opentelemetry-instrumentation-logging~=0.60b0
pexpect~=4.8
prometheus-client~=0.21.1
psutil~=7.0
pycares~=4.11.0
pycryptodome>=3.20.0
habana-pyhlml~=1.22.1.6
pyhumps~=3.8.0
pyroscope-io~=0.8.8
python-dateutil>=2.9
python-dotenv~=0.20.0
python-json-logger~=3.2.0
pyzmq~=26.4
PyJWT~=2.10.1
PyYAML~=6.0
pydantic[email]~=2.11.3
packaging>=24.1
hiredis>=3.0.0
huggingface-hub~=0.34.3
redis[hiredis]~=7.1.0
rich~=13.6
ruamel.yaml~=0.18.10
SQLAlchemy[postgresql_asyncpg]~=2.0.45
setproctitle~=1.3.5
setuptools~=80.0.0
strawberry-graphql~=0.278.0
tabulate~=0.8.9
temporenc~=0.1.0
tenacity>=9.0
toml~=0.10.2
tomli~=2.0.1
tomli-w~=1.2.0
tomlkit~=0.13.2
tqdm~=4.67.1
trafaret~=2.1
treelib~=1.7.0
tt-smi~=3.0.36; sys_platform == 'linux' and platform_machine == 'x86_64'
typeguard~=4.3
typing_extensions~=4.11
textual~=0.79.1
uvloop~=0.22.1; sys_platform != "Windows"
valkey-glide~=2.2.2
yarl~=1.19.0
zipstream-new~=1.1.8

# required by ai.backend.test (integration test suite)
pytest>=8.3.3
pytest-aiohttp~=1.1.0
pytest-dependency>=0.6.0
pytest-mock>=3.14.0
testcontainers[postgres,redis,minio]~=4.13.3

# type stubs
types-six
types-setuptools
types-python-dateutil
types-aiofiles
types-cachetools
types-colorama
types-Jinja2
# Removed types-networkx because it depends on numpy, which we do not want to install by default
# https://github.com/python/typeshed/issues/14760
# types-networkx
types-pexpect
types-psutil
types-PyYAML
types-redis
types-tabulate
types-toml
types-tqdm
backend.ai-krunner-alpine==5.4.0
backend.ai-krunner-static-gnu==4.4.0

etcd-client-py~=0.5.1
